Address review: move signature helpers to gmail_helpers.py, drop repro script
Per maintainer review on the PR:
- Move _HTMLSignatureExtractor and _signature_html_to_text from
  gmail_tools.py into gmail_helpers.py alongside the other helpers,
  importing _signature_html_to_text back into gmail_tools.py at its
  two call sites. The function's parameter is renamed html ->
  signature_html so it does not shadow the html module that
  gmail_helpers.py imports at top level.
- Remove scripts/repro_bugs.py, the one-off live-account harness used
  to reproduce the original bugs.

All 143 gmail tests still pass.
